Simplify emulator mcp code and address comments
