+++ src/mcp/tool.ts
@@ -4,7 +4,6 @@ import { zodToJsonSchema } from "zod-to-json-schema";
 import type { FirebaseMcpServer } from "./index";
 import { Config } from "../config";
 import { RC } from "../rc";
-import { EmulatorHubClient } from "../emulator/hubClient";
 
 export interface ServerToolContext {
   projectId?: string;
@@ -12,7 +11,6 @@ export interface ServerToolContext {
   config: Config;
   host: FirebaseMcpServer;
   rc: RC;
-  emulatorHubClient?: EmulatorHubClient;
 }
 
 export interface ServerTool<InputSchema extends ZodTypeAny = ZodTypeAny> {
