+++ src/mcp/tools/dataconnect/execute_graphql_read.ts
@@ -4,9 +4,8 @@ import { tool } from "../../tool.js";
 import * as dataplane from "../../../dataconnect/dataplaneClient.js";
 import { pickService } from "../../../dataconnect/fileUtils.js";
 import { graphqlResponseToToolResponse, parseVariables } from "./converter.js";
-import { mcpError } from "../../util.js";
 import { Client } from "../../../apiv2.js";
-import { getDataConnectEmulatorDetails } from "../../emulator/dataconnectEmulatorController.js";
+import { getDataConnectEmulatorClient } from "./emulator.js";
 
 export const execute_graphql_read = tool(
   {
@@ -40,29 +39,14 @@ export const execute_graphql_read = tool(
   },
   async (
     { query, serviceId, variables: unparsedVariables, useEmulator },
-    { projectId, config, emulatorHubClient },
+    { projectId, config, host },
   ) => {
-    if (!projectId) {
-      return mcpError("Project ID is required but not found.", "PROJECT_ID_MISSING");
-    }
-    const serviceInfo = await pickService(projectId, config, serviceId || undefined);
+    const serviceInfo = await pickService(projectId!, config, serviceId || undefined);
 
     let apiClient: Client;
 
     if (useEmulator) {
-      const emulatorDetails = await getDataConnectEmulatorDetails(emulatorHubClient);
-
-      if (!emulatorDetails) {
-        return mcpError(
-          "DataConnect emulator requested but not found or not running. Please ensure the emulator is started and the project ('firebase.json') is correctly configured.",
-          "EMULATOR_NOT_FOUND",
-        );
-      }
-      apiClient = new Client({
-        urlPrefix: emulatorDetails.url,
-        apiVersion: dataplane.DATACONNECT_API_VERSION,
-        auth: false,
-      });
+      apiClient = await getDataConnectEmulatorClient(await host.getEmulatorHubClient());
     } else {
       apiClient = dataplane.dataconnectDataplaneClient();
     }
